# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "description": "业务数据",
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse",
        "type": "object"
      },
...
      "CaptureFramePayload": {
        "properties": {
          "captured_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At Utc"
          },
          "frame_index": {
            "minimum": 0.0,
            "title": "Frame Index",
            "type": "integer"
          },
          "height": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "relative_path": {
            "title": "Relative Path",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "width": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "frame_index",
          "relative_path"
        ],
        "title": "CaptureFramePayload",
        "type": "object"
      },
      "CaptureSyncRequest": {
        "properties": {
          "error_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Reason"
          },
          "frames": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CaptureFramePayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frames"
          },
          "saved_frames": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved Frames"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "CaptureSyncRequest",
        "type": "object"
      },
...
      "CreateCaptureRequest": {
        "properties": {
          "gateway_id": {
            "title": "Gateway Id",
            "type": "string"
          },
          "max_frames": {
            "default": 300,
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Max Frames",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "sample_fps": {
            "default": 2.0,
            "exclusiveMinimum": 0.0,
            "maximum": 30.0,
            "title": "Sample Fps",
            "type": "number"
          },
          "save_dir": {
            "title": "Save Dir",
            "type": "string"
          },
          "save_format": {
            "default": "jpg",
            "title": "Save Format",
            "type": "string"
          },
          "source": {
            "default": "hdmi_x1301",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "gateway_id",
          "save_dir"
        ],
        "title": "CreateCaptureRequest",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "错误代码",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "description": "错误说明",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WeatherPayload": {
        "properties": {
          "cloudiness": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloudiness"
          },
          "fog_density": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fog Density"
          },
          "precipitation": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation"
          },
          "precipitation_deposits": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation Deposits"
          },
          "preset": {
            "default": "ClearNoon",
            "title": "Preset",
            "type": "string"
          },
          "sun_altitude_angle": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sun Altitude Angle"
          },
          "sun_azimuth_angle": {
            "anyOf": [
              {
                "maximum": 360.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sun Azimuth Angle"
          },
          "wetness": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wetness"
          },
          "wind_intensity": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Intensity"
          }
        },
        "title": "WeatherPayload",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "SimChip Nexus 芯境智测平台提供 CARLA 场景仿真、硬件在环测试、设备接入和评测报告服务。\n\n- `/docs`：接口文档\n- `/`：平台展示页\n- `/ui`：Web 测评控制台",
    "title": "SimChip Nexus API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/captures": {
      "get": {
        "operationId": "list_captures_captures_get",
        "parameters": [
          {
            "description": "可选采集状态过滤值",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "可选采集状态过滤值",
              "title": "Status"
            }
          },
          {
            "description": "可选网关过滤值",
            "in": "query",
            "name": "gateway_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "可选网关过滤值",
              "title": "Gateway Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询采集列表",
        "tags": [
          "采集管理"
        ]
      },
      "post": {
        "operationId": "create_capture_captures_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaptureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "创建采集任务",
        "tags": [
          "采集管理"
        ]
      }
    },
    "/captures/{capture_id}": {
      "get": {
        "operationId": "get_capture_captures__capture_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询单个采集",
        "tags": [
          "采集管理"
        ]
      }
    },
    "/captures/{capture_id}/frames": {
      "get": {
        "operationId": "get_capture_frames_captures__capture_id__frames_get",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "title": "Capture Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询采集帧列表",
        "tags": [
          "采集管理"
        ]
      }
    },
    "/captures/{capture_id}/manifest": {
      "get": {
        "operationId": "get_capture_manifest_captures__capture_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询采集 manifest",
        "tags": [
          "采集管理"
        ]
      }
    },
    "/captures/{capture_id}/start": {
      "post": {
        "operationId": "start_capture_captures__capture_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "启动采集",
        "tags": [
          "采集管理"
        ]
      }
    },
    "/captures/{capture_id}/stop": {
      "post": {
        "operationId": "stop_capture_captures__capture_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "停止采集",
        "tags": [
          "采集管理"
        ]
      }
    },
    "/captures/{capture_id}/sync": {
      "post": {
        "operationId": "sync_capture_captures__capture_id__sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "同步采集进度",
        "tags": [
          "采集管理"
        ]
      }
    },
...
    "/system/carla/weather": {
      "put": {
        "operationId": "update_carla_weather_system_carla_weather_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "动态修改全局天气",
        "tags": [
          "全局仿真控制"
        ]
      }
    },
...
  }
}